+++ src/swift/Sources/APIViewManager.swift
@@ -38,7 +38,7 @@ class APIViewManager {
     let args = CommandLineArguments()
     var tokenFile = TokenFile(name: "TestFile")
 
-    let publicModifiers: [AccessLevelModifier] = [.public, .open]
+    
 
     // MARK: Methods
 
@@ -82,174 +82,8 @@ class APIViewManager {
         let sourceFile = try SourceReader.read(at: sourceUrl.absoluteString)
         let parser = Parser(source: sourceFile)
         let topLevelDecl = try parser.parse()
-        process(topLevelDecl)
+        tokenFile.process(topLevelDecl)
     }
 
-    func process(_ decl: TopLevelDeclaration) {
-        for statement in decl.statements {
-            switch statement {
-            case let decl as ClassDeclaration:
-              process(decl)
-            case let decl as ConstantDeclaration:
-              process(decl)
-            case let decl as DeinitializerDeclaration:
-              process(decl)
-            case let decl as EnumDeclaration:
-              process(decl)
-            case let decl as ExtensionDeclaration:
-              process(decl)
-            case let decl as FunctionDeclaration:
-              process(decl)
-            case let decl as ImportDeclaration:
-              process(decl)
-            case let decl as InitializerDeclaration:
-              process(decl)
-            case let decl as OperatorDeclaration:
-              continue // process(decl)
-            case let decl as PrecedenceGroupDeclaration:
-              continue // process(decl)
-            case let decl as ProtocolDeclaration:
-              process(decl)
-            case let decl as StructDeclaration:
-              process(decl)
-            case let decl as SubscriptDeclaration:
-              continue // process(decl)
-            case let decl as TypealiasDeclaration:
-              process(decl)
-            case let decl as VariableDeclaration:
-              process(decl)
-            default:
-                continue
-            }
-        }
-    }
-
-    func process(_ decl: ClassDeclaration) {
-        guard publicModifiers.contains(decl.accessLevelModifier ?? .internal) else {
-            return
-        }
-        let value = (decl.accessLevelModifier ?? .internal).textDescription
-        tokenFile.addKeyword(value: value)
-        tokenFile.addWhitespace()
-        if decl.isFinal {
-            tokenFile.addKeyword(value: "final")
-            tokenFile.addWhitespace()
-        }
-        tokenFile.addType(name: decl.name.textDescription)
-        if let inheritance = decl.typeInheritanceClause {
-            process(inheritance)
-        }
-        tokenFile.addPunctuation("{")
-        tokenFile.indentLevel += 2
-        tokenFile.addNewline()
-        for member in decl.members {
-            // TODO: Add members
-            switch member {
-            case .declaration(let decl):
-                process(decl)
-            default:
-                continue
-            }
-        }
-        tokenFile.indentLevel -= 2
-        tokenFile.addNewline()
-        tokenFile.addPunctuation("}")
-        tokenFile.addNewline()
-    }
-
-    func process(_ decl: StructDeclaration) {
-        guard publicModifiers.contains(decl.accessLevelModifier ?? .internal) else {
-            return
-        }
-    }
-
-    func process(_ decl: EnumDeclaration) {
-        guard publicModifiers.contains(decl.accessLevelModifier ?? .internal) else {
-            return
-        }
-    }
-
-    func process(_ decl: ProtocolDeclaration) {
-        
-    }
     
-    func process(_ decl: TypealiasDeclaration) {
-        
-    }
-    
-    func process(_ decl: VariableDeclaration) {
-        
-    }
-    
-    func process(_ decl: ExtensionDeclaration) {
-        
-    }
-    
-    func process(_ decl: ConstantDeclaration) {
-        
-    }
-    
-    func process(_ decl: InitializerDeclaration) {
-        
-    }
-    
-    func process(_ decl: DeinitializerDeclaration) {
-        
-    }
-    
-    func process(_ decl: FunctionDeclaration) {
-        
-    }
-    
-    func process(_ decl: ImportDeclaration) {
-        
-    }
-    
-
-    
-    func process(_ clause: TypeInheritanceClause) {
-        tokenFile.addPunctuation(":")
-        tokenFile.addWhitespace()
-        for item in clause.typeInheritanceList {
-            // TODO: Add type inheritance
-            let test = "best"
-        }
-    }
-    
-    func process(_ decl: Declaration) {
-        switch decl {
-        case let decl as ClassDeclaration:
-          return process(decl)
-        case let decl as ConstantDeclaration:
-          return process(decl)
-        case let decl as DeinitializerDeclaration:
-          return process(decl)
-        case let decl as EnumDeclaration:
-          return process(decl)
-        case let decl as ExtensionDeclaration:
-          return process(decl)
-        case let decl as FunctionDeclaration:
-          return process(decl)
-        case let decl as ImportDeclaration:
-          return process(decl)
-        case let decl as InitializerDeclaration:
-          return process(decl)
-        case let decl as OperatorDeclaration:
-          return // process(decl)
-        case let decl as PrecedenceGroupDeclaration:
-          return // process(decl)
-        case let decl as ProtocolDeclaration:
-          return process(decl)
-        case let decl as StructDeclaration:
-          return process(decl)
-        case let decl as SubscriptDeclaration:
-          return // process(decl)
-        case let decl as TypealiasDeclaration:
-          return process(decl)
-        case let decl as VariableDeclaration:
-          return process(decl)
-        default:
-          return // no implementation for this declaration, just continue
-        }
-    }
 }
